Cleanup SettingsManager, FontManager, SettingsPanel
-Removed redundant functions, members, function parameters.
-Moved DisplaySettings dpi out of json.
-Moved DisplaySettings into new UserSettings in preparation for multiple categories of settings.
-Simplified flow for modifying user settings: Caller directly modifies settings kept by SettingsManager, then calls ApplyUserSettings().
-Change themes by altering ptr instead of copying color array.
-New settings panel.
-Block inputs to timeline when popups are open.

+++ src/view/src/rocprofvis_events_view.cpp
@@ -212,12 +212,11 @@ bool
 EventsView::XButton()
 {
     bool clicked = false;
-    ImGui::PushStyleColor(ImGuiCol_Button,
-                          m_settings.GetColor(static_cast<int>(Colors::kTransparent)));
+    ImGui::PushStyleColor(ImGuiCol_Button, m_settings.GetColor(Colors::kTransparent));
     ImGui::PushStyleColor(ImGuiCol_ButtonHovered,
-                          m_settings.GetColor(static_cast<int>(Colors::kTransparent)));
+                          m_settings.GetColor(Colors::kTransparent));
     ImGui::PushStyleColor(ImGuiCol_ButtonActive,
-                          m_settings.GetColor(static_cast<int>(Colors::kTransparent)));
+                          m_settings.GetColor(Colors::kTransparent));
     ImGui::PushFont(m_settings.GetFontManager().GetIconFont(FontType::kDefault));
     clicked = ImGui::SmallButton(ICON_X_CIRCLED);
     ImGui::PopFont();
