+++ src/view/src/widgets/rocprofvis_widget.cpp
@@ -135,9 +135,8 @@ VFixedContainer::Render()
 
         ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, m_children[i].m_item_spacing);
         ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, m_children[i].m_window_padding);
-        ImGui::PushStyleColor(
-            ImGuiCol_ChildBg,
-            SettingsManager::GetInstance().GetColor(static_cast<int>(Colors::kFillerColor)));
+        ImGui::PushStyleColor(ImGuiCol_ChildBg, SettingsManager::GetInstance().GetColor(
+                                                    Colors::kFillerColor));
 
         ImGui::BeginChild(ImGui::GetID(static_cast<int>(i)),
                           ImVec2(m_children[i].m_width, m_children[i].m_height),
@@ -215,8 +214,8 @@ HSplitContainer::Render()
     ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, m_left.m_item_spacing);
     ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, m_left.m_window_padding);
 
-    ImGui::PushStyleColor(ImGuiCol_ChildBg, SettingsManager::GetInstance().GetColor(
-                                                static_cast<int>(Colors::kFillerColor)));
+    ImGui::PushStyleColor(ImGuiCol_ChildBg,
+                          SettingsManager::GetInstance().GetColor(Colors::kFillerColor));
     ImGui::BeginChild(m_left_name.c_str(), ImVec2(left_col_width, col_height),
                       m_left.m_child_flags, m_left.m_window_flags);
     if(m_left.m_item)
@@ -239,7 +238,7 @@ HSplitContainer::Render()
     ImVec2 splitter_max = ImGui::GetItemRectMax();
     ImGui::GetWindowDrawList()->AddRectFilled(
         splitter_min, splitter_max,
-        SettingsManager::GetInstance().GetColor(static_cast<int>(Colors::kSplitterColor)));
+        SettingsManager::GetInstance().GetColor(Colors::kSplitterColor));
 
     // Change cursor appearance when hovering over the resize handle
     if(ImGui::IsItemHovered())
@@ -264,10 +263,10 @@ HSplitContainer::Render()
     ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, m_right.m_item_spacing);
     ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, m_right.m_window_padding);
 
-    ImGui::PushStyleColor(ImGuiCol_ChildBg, SettingsManager::GetInstance().GetColor(
-                                                static_cast<int>(Colors::kFillerColor)));
-    ImGui::BeginChild(m_right_name.c_str(), ImVec2(-1, col_height),
-                      m_right.m_child_flags, m_right.m_window_flags);
+    ImGui::PushStyleColor(ImGuiCol_ChildBg,
+                          SettingsManager::GetInstance().GetColor(Colors::kFillerColor));
+    ImGui::BeginChild(m_right_name.c_str(), ImVec2(-1, col_height), m_right.m_child_flags,
+                      m_right.m_window_flags);
     if(m_right.m_item)
     {
         m_right.m_item->Render();
@@ -339,8 +338,8 @@ VSplitContainer::Render()
     // Start Top Row
     ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, m_top.m_item_spacing);
     ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, m_top.m_window_padding);
-    ImGui::PushStyleColor(ImGuiCol_ChildBg, SettingsManager::GetInstance().GetColor(
-                                                static_cast<int>(Colors::kFillerColor)));
+    ImGui::PushStyleColor(ImGuiCol_ChildBg,
+                          SettingsManager::GetInstance().GetColor(Colors::kFillerColor));
     ImGui::BeginChild(m_top_name.c_str(), ImVec2(col_width, top_row_height),
                       m_top.m_child_flags, m_top.m_window_flags);
     if(m_top.m_item)
@@ -360,7 +359,7 @@ VSplitContainer::Render()
     ImVec2 splitter_max = ImGui::GetItemRectMax();
     ImGui::GetWindowDrawList()->AddRectFilled(
         splitter_min, splitter_max,
-        SettingsManager::GetInstance().GetColor(static_cast<int>(Colors::kSplitterColor)));
+        SettingsManager::GetInstance().GetColor(Colors::kSplitterColor));
 
     // Change cursor appearance when hovering over the resize handle
     if(ImGui::IsItemHovered())
@@ -383,10 +382,10 @@ VSplitContainer::Render()
     ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, m_bottom.m_item_spacing);
     ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, m_bottom.m_window_padding);
 
-    ImGui::PushStyleColor(ImGuiCol_ChildBg, SettingsManager::GetInstance().GetColor(
-                                                static_cast<int>(Colors::kFillerColor)));
-    ImGui::BeginChild(m_bottom_name.c_str(), ImVec2(col_width, 0),
-                      m_bottom.m_child_flags, m_bottom.m_window_flags);
+    ImGui::PushStyleColor(ImGuiCol_ChildBg,
+                          SettingsManager::GetInstance().GetColor(Colors::kFillerColor));
+    ImGui::BeginChild(m_bottom_name.c_str(), ImVec2(col_width, 0), m_bottom.m_child_flags,
+                      m_bottom.m_window_flags);
     if(m_bottom.m_item)
     {
         m_bottom.m_item->Render();
